requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
boto3==1.36.23
botocore==1.36.23
click==8.1.8
colorama==0.4.6
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.8
h11==0.14.0
idna==3.10
jmespath==1.0.1
passlib==1.7.4
pyasn1==0.4.8
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9
s3transfer==0.11.2
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.34.0